ad7091r: Rename files
